pyproject.toml
[tool.poetry]
name = "lw1"
version = "0.1.0"
description = ""
authors = ["Lukas Winkler <git@lw1.at>"]

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^3.1.1"
toml = "^0.10.2"
watchfiles = "^0.13"
python-slugify = "^6.1.1"
PyYAML = "^6.0"
Markdown = "^3.3.6"
blurhash-python = "^1.1.3"
Pygments = "^2.11.2"
markdown-katex = "^202112.1034"
Babel = "^2.10.1"
redis = "^4.2.2"
hiredis = "^2.0.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
lw1 = 'lw1.main:main'


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

